Update Content Provider MyNote

+++ ContentProvider/MyNotesApp/java/app/src/main/java/com/dicoding/picodiploma/mynotesapp/provider/NoteProvider.java
@@ -83,13 +83,10 @@ public class NoteProvider extends ContentProvider {
     @Override
     public Uri insert(@NonNull Uri uri, ContentValues contentValues) {
         long added;
-        switch (sUriMatcher.match(uri)) {
-            case NOTE:
-                added = noteHelper.insert(contentValues);
-                break;
-            default:
-                added = 0;
-                break;
+        if (sUriMatcher.match(uri) == NOTE) {
+            added = noteHelper.insert(contentValues);
+        } else {
+            added = 0;
         }
 
         getContext().getContentResolver().notifyChange(CONTENT_URI, null);
@@ -101,13 +98,10 @@ public class NoteProvider extends ContentProvider {
     @Override
     public int update(@NonNull Uri uri, ContentValues contentValues, String s, String[] strings) {
         int updated;
-        switch (sUriMatcher.match(uri)) {
-            case NOTE_ID:
-                updated = noteHelper.update(uri.getLastPathSegment(), contentValues);
-                break;
-            default:
-                updated = 0;
-                break;
+        if (sUriMatcher.match(uri) == NOTE_ID) {
+            updated = noteHelper.update(uri.getLastPathSegment(), contentValues);
+        } else {
+            updated = 0;
         }
 
         getContext().getContentResolver().notifyChange(CONTENT_URI, null);
@@ -118,13 +112,10 @@ public class NoteProvider extends ContentProvider {
     @Override
     public int delete(@NonNull Uri uri, String s, String[] strings) {
         int deleted;
-        switch (sUriMatcher.match(uri)) {
-            case NOTE_ID:
-                deleted = noteHelper.deleteById(uri.getLastPathSegment());
-                break;
-            default:
-                deleted = 0;
-                break;
+        if (sUriMatcher.match(uri) == NOTE_ID) {
+            deleted = noteHelper.deleteById(uri.getLastPathSegment());
+        } else {
+            deleted = 0;
         }
 
         getContext().getContentResolver().notifyChange(CONTENT_URI, null);
